desktop/requirements-desktop.txt
# Reuse the main project requirements plus desktop-only helpers.
-r ../requirements.txt

# Desktop shell
pywebview==4.4.1
pyinstaller==6.3.0
pyyaml==6.0.1
# pkg_resources runtime deps to avoid missing jaraco errors in frozen build
jaraco.functools==3.7.0
jaraco.context==4.3.0
jaraco.text==3.11.1
